pyproject.toml
[project]
name = "jeroenoverschie-nl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "xmltodict>=0.14.2",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "pytest>=8.3.3",
    "ruff>=0.7.1",
]
